Removed a bunch of compile warnings fucked up the text rendering.
Former-commit-id: f3ae5c546603860097566e930760d33160787e00

+++ CommonLib/DataStructures/Hashmap/Hash.cpp
@@ -332,7 +332,7 @@ void MurmurHash3_x64_128(const void * key, const int len, const uint32_t seed, v
 
 u32 Hash(const char* key)
 {
-	s32 size = strlen(key);
+	s32 size = (s32)strlen(key);
 	u32 result;
 	MurmurHash3_x86_32(key, size, 0, &result);
 	return result;
